requirements.txt
transformers==4.57.0
huggingface-hub==0.35.3
tokenizers==0.22.1
safetensors==0.6.2
numpy>=1.24,<2.4
opencv-python==4.12.0.88
Pillow==11.3.0
torch==2.8.0
Levenshtein==0.27.1
rapidfuzz==3.14.1
httpx>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
jinja2>=3.0.0
easyocr>=1.6.0
pyyaml>=6.0.0
regex>=2025.0.0
filelock
fsspec